requirements.txt
# Core ML/AI Libraries
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0
accelerate>=0.24.0
peft>=0.7.0  # Parameter-Efficient Fine-Tuning (LoRA, etc.)
bitsandbytes>=0.41.0  # For quantization
trl>=0.7.0  # Transformer Reinforcement Learning

# Data Processing
pandas>=2.0.0
numpy>=1.24.0
jsonlines>=4.0.0
pyyaml>=6.0

# API and Serving
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.4.0
python-multipart>=0.0.6

# Database and Authentication
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
email-validator>=2.0.0

# Evaluation and Metrics
scikit-learn>=1.3.0
rouge-score>=0.1.2
nltk>=3.8.1

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
colorama>=0.4.6
click>=8.1.0

# Development and Testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.0

# Logging and Monitoring
loguru>=0.7.0
wandb>=0.16.0  # Optional: for experiment tracking
psutil>=5.9.0  # System monitoring for admin dashboard

# Optional: OpenAI API for comparison/baseline
openai>=1.3.0
anthropic>=0.7.0
